Switch index from u64 to u128 [13 tests currently failing]

+++ src/ownedchain.rs
@@ -140,7 +140,7 @@ impl OwnedChain {
     }
 
     /// Number of blocks in this owned chain.
-    pub fn count(&self) -> u64 {
+    pub fn count(&self) -> u128 {
         self.chain.count()
     }
